feat(eternal-jukebox): Start visualizer refactor
