+++ custom-apps/eternal-jukebox/src/components/visualizer-slice.component.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import styles from '../css/app.module.scss';
 import { IBeatDrawData } from '../models/visualization/beat-draw-data.interface';
 
@@ -6,35 +6,22 @@ interface IProps {
     drawData: IBeatDrawData;
 }
 
-interface IState {
-    isHovered: boolean;
-}
+export function VisualizerSlice(props: IProps) {
+    const [isHovered, setIsHovered] = useState(false);
 
-export class VisualizerSlice extends React.Component<IProps, IState> {
-    constructor(props: IProps) {
-        super(props);
-
-        this.state = {
-            isHovered: false,
-        };
-    }
-
-    // TODO: Seek on click
-    render() {
-        return (
-            <path
-                className={styles['beat-path']}
-                fill={
-                    this.props.drawData.beat.isPlaying || this.state.isHovered
-                        ? this.props.drawData.activeColor
-                        : this.props.drawData.color
-                }
-                d={this.props.drawData.drawCommand}
-                onMouseOver={() => this.setState(() => ({ isHovered: true }))}
-                onMouseOut={() => this.setState(() => ({ isHovered: false }))}
-            >
-                <title>Beat {this.props.drawData.beat.index}</title>
-            </path>
-        );
-    }
+    return (
+        <path
+            className={styles['beat-path']}
+            fill={
+                props.drawData.beat.isPlaying || isHovered
+                    ? props.drawData.activeColor
+                    : props.drawData.color
+            }
+            d={props.drawData.drawCommand}
+            onMouseOver={() => setIsHovered(true)}
+            onMouseOut={() => setIsHovered(false)}
+        >
+            <title>Beat {props.drawData.beat.index}</title>
+        </path>
+    );
 }
