+++ custom-apps/eternal-jukebox/src/models/graph/graph-state.ts
@@ -1,0 +1,8 @@
+import { RemixedSegment, RemixedTimeInterval } from '../remixer.types';
+import { Beat } from './beat';
+
+export interface IGraphState {
+    beats: Beat[];
+    segments: RemixedSegment[];
+    remixedBeats: RemixedTimeInterval[];
+}
