+++ custom-apps/eternal-jukebox/src/models/visualization/graph-draw-data.ts
@@ -1,0 +1,7 @@
+import { IBeatDrawData } from './beat-draw-data.interface';
+import { IEdgeDrawData } from './edge-draw-data.interface';
+
+export interface IGraphDrawData {
+    beats: IBeatDrawData[];
+    edges: IEdgeDrawData[];
+}
